Functional LED Class
LED class now uses fully templated class to configure the GPIO pin and
present the program with on, off and toggle options.

The only remaining issue is the setting of the pin mode type currently
requires a hard coded value.  This should use a constexpr but could also
use the XMC_Lib #define values.