First Functional Blinky
The CCU and CC have been configured to issue a periodic ISR that toggles
the LED on the board.  Works correctly.

TODO:
Remove calls to the XMCLib - Added to quickly interface with the NVIC
module rather than create an NVIC class yet.

Clean up the GPT and move most of the CCU/CC configurations to the CCU4
class.  The GPT class should only do things directly related to the
timer function.  The CCU class should handle configuring the CCU and a
CC class should handle configuring the CC sub module.